removed "web" package as don't have any api endpoint. (#34)
