+++ domain-services/fiscal-event-post-processor/CODECOVERAGEREPORT.md
@@ -1,5 +1,5 @@
 # Junit Code Coverage Report
 All junit code coverage report to this module will be documented in this file.
 
-## 1.0.2 - 2021-12-06
-- Code coverage report  ![fiscal-event-post-processor-junit-report-1.0.2.png](fiscal-event-post-processor-junit-report-1.0.2.png)
+## 1.0.4 - 2022-01-28
+- Code coverage report  ![fiscal-event-post-processor-junit-report-1.0.4.png](fiscal-event-post-processor-junit-report-1.0.4.png)
