+++ domain-services/fiscal-event-post-processor/fiscal-event-post-processor-junit-report-1.0.4.png
