+++ domain-services/fiscal-event-post-processor/pom.xml
@@ -5,7 +5,7 @@
   <artifactId>fiscal-event-post-processor</artifactId>
   <packaging>jar</packaging>
   <name>fiscal-event-post-processor</name>
-  <version>1.0.3</version>
+  <version>1.0.4</version>
   <properties>
     <log4j2.version>2.17.1</log4j2.version>
     <java.version>1.8</java.version>
