+++ domain-services/fiscal-event-post-processor/src/main/java/org/egov/consumer/FiscalEventDereferenceConsumer.java
@@ -6,8 +6,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.egov.config.FiscalEventPostProcessorConfig;
 import org.egov.producer.Producer;
 import org.egov.service.FiscalEventDereferenceService;
-import org.egov.web.models.FiscalEventDeReferenced;
-import org.egov.web.models.FiscalEventRequest;
+import org.egov.models.FiscalEventDeReferenced;
+import org.egov.models.FiscalEventRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.kafka.support.KafkaHeaders;
