+++ domain-services/fiscal-event-post-processor/src/main/java/org/egov/consumer/FiscalEventUnbundledFlattenConsumer.java
@@ -8,8 +8,8 @@ import org.egov.config.FiscalEventPostProcessorConfig;
 import org.egov.producer.Producer;
 import org.egov.service.FiscalEventFlattenService;
 import org.egov.service.FiscalEventUnbundleService;
-import org.egov.web.models.FiscalEventDeReferenced;
-import org.egov.web.models.FiscalEventLineItemUnbundled;
+import org.egov.models.FiscalEventDeReferenced;
+import org.egov.models.FiscalEventLineItemUnbundled;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.kafka.support.KafkaHeaders;
