+++ domain-services/fiscal-event-post-processor/src/main/java/org/egov/models/FiscalEventDruid.java
@@ -1,4 +1,4 @@
-package org.egov.web.models;
+package org.egov.models;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
