+++ domain-services/fiscal-event-post-processor/src/main/java/org/egov/service/FiscalEventDereferenceService.java
@@ -4,8 +4,8 @@ package org.egov.service;
 import com.fasterxml.jackson.databind.JsonNode;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
+import org.egov.models.*;
 import org.egov.util.*;
-import org.egov.web.models.*;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
