+++ domain-services/fiscal-event-post-processor/src/main/java/org/egov/service/FiscalEventFlattenService.java
@@ -3,7 +3,7 @@ package org.egov.service;
 import com.github.wnameless.json.flattener.JsonFlattener;
 import com.google.gson.Gson;
 import lombok.extern.slf4j.Slf4j;
-import org.egov.web.models.FiscalEventLineItemUnbundled;
+import org.egov.models.FiscalEventLineItemUnbundled;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
