+++ domain-services/fiscal-event-post-processor/src/main/java/org/egov/util/CoaUtil.java
@@ -10,9 +10,9 @@ import org.egov.common.contract.request.RequestHeader;
 import org.egov.config.FiscalEventPostProcessorConfig;
 import org.egov.resposioty.ServiceRequestRepository;
 import org.egov.tracer.model.CustomException;
-import org.egov.web.models.Amount;
-import org.egov.web.models.ChartOfAccount;
-import org.egov.web.models.FiscalEvent;
+import org.egov.models.Amount;
+import org.egov.models.ChartOfAccount;
+import org.egov.models.FiscalEvent;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
