+++ domain-services/fiscal-event-post-processor/src/test/java/org/egov/consumer/FiscalEventDereferenceConsumerTest.java
@@ -5,8 +5,8 @@ import org.egov.config.FiscalEventPostProcessorConfig;
 import org.egov.config.TestDataFormatter;
 import org.egov.producer.Producer;
 import org.egov.service.FiscalEventDereferenceService;
-import org.egov.web.models.FiscalEventDeReferenced;
-import org.egov.web.models.FiscalEventRequest;
+import org.egov.models.FiscalEventDeReferenced;
+import org.egov.models.FiscalEventRequest;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
