+++ domain-services/fiscal-event-post-processor/src/test/java/org/egov/service/FiscalEventDereferenceEnrichmentServiceTest.java
@@ -1,8 +1,8 @@
 package org.egov.service;
 
 import org.egov.config.TestDataFormatter;
-import org.egov.web.models.FiscalEventDeReferenced;
-import org.egov.web.models.FiscalEventRequest;
+import org.egov.models.FiscalEventDeReferenced;
+import org.egov.models.FiscalEventRequest;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
