+++ domain-services/fiscal-event-post-processor/src/test/java/org/egov/service/FiscalEventUnbundleServiceTest.java
@@ -1,7 +1,7 @@
 package org.egov.service;
 
 import org.egov.config.TestDataFormatter;
-import org.egov.web.models.FiscalEventDeReferenced;
+import org.egov.models.FiscalEventDeReferenced;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
