+++ domain-services/fiscal-event-post-processor/src/test/java/org/egov/util/ExpenditureUtilTest.java
@@ -8,8 +8,8 @@ import org.egov.config.FiscalEventPostProcessorConfig;
 import org.egov.config.TestDataFormatter;
 import org.egov.resposioty.ServiceRequestRepository;
 import org.egov.tracer.model.CustomException;
-import org.egov.web.models.Expenditure;
-import org.egov.web.models.FiscalEventRequest;
+import org.egov.models.Expenditure;
+import org.egov.models.FiscalEventRequest;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
