+++ domain-services/fiscal-event-post-processor/src/test/java/org/egov/util/GovernmentUtilTest.java
@@ -8,9 +8,9 @@ import org.egov.config.FiscalEventPostProcessorConfig;
 import org.egov.config.TestDataFormatter;
 import org.egov.resposioty.ServiceRequestRepository;
 import org.egov.tracer.model.CustomException;
-import org.egov.web.models.FiscalEvent;
-import org.egov.web.models.FiscalEventRequest;
-import org.egov.web.models.Government;
+import org.egov.models.FiscalEvent;
+import org.egov.models.FiscalEventRequest;
+import org.egov.models.Government;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
